fix types to work with noImplicitAny
no implict any is very commonly used in TS projects. this project previously could not compile when imported under those circumstances. This PR also includes formatting changes that are standard for TS but if that is a problem we could try to not do those.. but preferably the types just use TS standard formatting.